excise ReposTableHeader, make sticky, refactor ColorCircle
